Refactored abbreviation and label creation
